Reparied testing scripts for jec17ul

+++ Skimming/zjet/configs/jec17ul/1060_DYJetsToLL_Summer19-madgraphMLM_mc2017_realistic_v6-v3/testKappaSkim.sh
@@ -2,20 +2,41 @@
 
 _GT="106X_mc2017_realistic_v6"
 _NEVT=100
-_IS_DATA=0
+_IS_DATA=false
 _GRID_PATH_PREFIX="root://xrootd-cms.infn.it/"
 _FILE="/store/mc/RunIISummer19UL17MiniAOD/DYJetsToLL_M-50_TuneCP5_13TeV-madgraphMLM-pythia8/MINIAODSIM/106X_mc2017_realistic_v6-v2/70000/FBC4E43E-2C35-974A-84C9-29AD0430BD39.root"
 _DIR="test/${_GT}"
 
-mkdir -p $_dir
+mkdir -p $_DIR
+
+echo "##############################################################################"
+echo "STARTING cmsRun !!!!"
+echo "##############################################################################"
 
 cmsRun ../kappaSkim_default_cfg.py inputFiles=${_GRID_PATH_PREFIX}${_FILE} \
                         globalTag=${_GT} \
-                        isData=_IS_DATA \
+                        isData=${_IS_DATA} \
                         outputFile=testKappaSkim_out_${_GT}.root \
                         maxEvents=${_NEVT} \
                         dumpPython=1 2>&1 | tee cout_${_GT}.log
-mv cout_$_GT.log $_DIR/
-mv infos.log $_DIR/
-mv debugs.log $_DIR/
-mv testKappaSkim_out_${_GT}_numEvent${_NEVT}.root $_DIR/
+
+echo "##############################################################################"
+echo "cmsRun terminated !!!!"
+echo "##############################################################################"
+
+if [ -f cout_$_GT.log ]; then
+    mv cout_$_GT.log $_DIR/
+fi
+
+if [ -f infos.log ]; then
+    mv infos.log $_DIR/
+fi
+
+if [ -f debugs.log ]; then
+    mv debugs.log $_DIR/
+fi
+
+if [ -f testKappaSkim_out_${_GT}_numEvent${_NEVT}.root ]; then
+    mv testKappaSkim_out_${_GT}_numEvent${_NEVT}.root $_DIR/
+fi
+
